+++ .flowconfig
@@ -14,3 +14,4 @@ suppress_comment= \\(.\\|\n\\)*\\$FlowFixMe
 esproposal.class_static_fields=enable
 esproposal.class_instance_fields=enable
 unsafe.enable_getters_and_setters=true
+module.name_mapper='/lib/' -> '/src/'
