+++ bin/prepack-repl.js
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+require("../lib/repl-cli");
