+++ circle.yml
@@ -17,8 +17,8 @@ test:
     - npm run test-test262 -- --statusFile $CIRCLE_ARTIFACTS/test262-status.txt --timeout 100 --cpuScale 0.25 --verbose:
         timeout: 1800
   post:
-    - mv lib/coverage/lcov-report $CIRCLE_ARTIFACTS/coverage-report
-    - mv lib/coverage-sourcemapped $CIRCLE_ARTIFACTS/coverage-report-sourcemapped
+    - mv coverage/lcov-report $CIRCLE_ARTIFACTS/coverage-report
+    - mv coverage-sourcemapped $CIRCLE_ARTIFACTS/coverage-report-sourcemapped
 
 checkout:
   post:
