+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "prepack",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "description": "Execute a JS bundle, serialize global state and side effects to a snapshot that can be quickly restored.",
   "homepage": "https://github.com/facebook/prepack",
   "repository": {
@@ -11,58 +11,55 @@
     "url": "https://github.com/facebook/prepack/issues"
   },
   "files": [
-    "src/*",
-    "lib/*"
+    "LICENSE",
+    "PATENTS",
+    "bin/",
+    "lib/"
   ],
+  "bin": {
+    "prepack": "bin/prepack.js",
+    "prepack-repl": "bin/prepack-repl.js"
+  },
   "main": "lib/prepack.js",
   "scripts": {
     "build": "babel src --out-dir lib --source-maps && npm run build-bundle",
     "build-bundle": "webpack",
     "watch": "babel src --out-dir lib --watch --source-maps",
-    "lint": "eslint src",
+    "lint": "eslint src scripts",
     "flow": "flow check",
-    "test-serializer": "node lib/scripts/test-runner.js",
-    "test-serializer-with-coverage": "cd lib && ../node_modules/.bin/istanbul cover scripts/test-runner.js && ../node_modules/.bin/remap-istanbul -i coverage/coverage.json -o coverage-sourcemapped -t html && echo You can find coverage reports here: && echo   lib/coverage/lcov-report/index.html && echo   lib/coverage-sourcemapped/index.html",
-    "test-sourcemaps": "node lib/scripts/generate-sourcemaps-test.js && bash < src/scripts/test-sourcemaps.sh",
-    "test-test262": "node lib/scripts/test262-runner.js",
-    "test-internal": "node --stack_size=10000 --stack_trace_limit=200 --max_old_space_size=8192 lib/scripts/test-internal.js",
+    "test-serializer": "node scripts/test-runner.js",
+    "test-serializer-with-coverage": "./node_modules/.bin/istanbul cover ./scripts/test-runner.js && ./node_modules/.bin/remap-istanbul -i coverage/coverage.json -o coverage-sourcemapped -t html && echo You can find coverage reports here: && echo   coverage/lcov-report/index.html && echo   coverage-sourcemapped/index.html",
+    "test-sourcemaps": "babel-node scripts/generate-sourcemaps-test.js && bash < scripts/test-sourcemaps.sh",
+    "test-test262": "babel-node scripts/test262-runner.js",
+    "test-internal": "babel-node --stack_size=10000 --stack_trace_limit=200 --max_old_space_size=8192 scripts/test-internal.js",
     "test": "npm run test-serializer && npm run test-sourcemaps && npm run test-test262 && npm run test-internal",
-    "repl": "node lib/repl.js",
-    "prepack": "node lib/run_util.js",
+    "repl": "node lib/repl-cli.js",
+    "prepack": "node lib/prepack-cli.js",
     "validate": "npm run build && npm run lint && npm run depcheck && npm run flow && npm test",
     "prepublish": "npm run build",
-    "depcheck": "node lib/scripts/detect_bad_deps.js"
+    "depcheck": "babel-node scripts/detect_bad_deps.js"
   },
   "dependencies": {
     "babel-core": "^6.8.0",
     "babel-generator": "^6.8.0",
-    "babel-helper-function-name": "^6.8.0",
-    "babel-helper-get-function-arity": "^6.8.0",
-    "babel-preset-react": "^6.5.0",
     "babel-template": "^6.9.0",
     "babel-traverse": "^6.9.0",
     "babel-types": "^6.9.0",
     "babylon": "^6.5.2",
     "base62": "^1.1.2",
-    "chalk": "^1.1.3",
-    "invariant": "^2.2.0",
-    "istanbul": "^0.4.5",
-    "js-yaml": "^3.6.1",
-    "jsdom": "^9.2.1",
-    "minimist": "^1.2.0",
-    "remap-istanbul": "^0.9.1",
-    "repeating": "^2.0.0",
     "seedrandom": "^2.4.2",
-    "source-map": "^0.5.6",
-    "source-map-support": "^0.4.6",
-    "uglify-js": "^2.6.2"
+    "source-map": "^0.5.6"
   },
   "devDependencies": {
     "babel-cli": "^6.18.0",
     "babel-eslint": "^7.1.1",
+    "babel-helper-function-name": "^6.8.0",
+    "babel-helper-get-function-arity": "^6.8.0",
     "babel-plugin-transform-class-properties": "^6.6.0",
     "babel-plugin-transform-object-rest-spread": "^6.8.0",
     "babel-preset-es2015": "^6.5.0",
+    "babel-preset-react": "^6.5.0",
+    "chalk": "^1.1.3",
     "eslint": "^3.8.1",
     "eslint-plugin-babel": "^3.3.0",
     "eslint-plugin-flow-header": "^0.1.1",
@@ -70,8 +67,16 @@
     "eslint-plugin-header": "^1.0.0",
     "flow-bin": "^0.44.0",
     "graceful-fs": "^4.1.11",
+    "invariant": "^2.2.0",
+    "istanbul": "^0.4.5",
+    "js-yaml": "^3.6.1",
+    "jsdom": "^9.2.1",
     "kcheck": "^2.0.0",
     "madge": "^1.6.0",
+    "minimist": "^1.2.0",
+    "remap-istanbul": "^0.9.1",
+    "source-map-support": "^0.4.6",
+    "uglify-js": "^2.6.2",
     "webpack": "^2.3.3"
   },
   "engines": {
