+++ scripts/detect_bad_deps.js
