+++ scripts/test-sourcemaps.sh
